Revert "Set height and width specifically on img tag"
This reverts commit be12afb50d56900356a5904ef0aeea7482274569.

It causes issues with aspect ratio when the raw width of the image is
wider than the width in CSS pixels allowed by the body max-width.

Not reverting the refactor that removes {{ $img = . }} because I don't
know why that was there and will see if it works fine with it still
removed.