Remove share buttons
About 0.19% of visitors in the past 12 months have clicked the share
buttons at all. The code is a little ugly, and it wastes a tiny bit of
space at the bottom of the page. I don't think it's really worth having.

And I think people probably print a page through the browser UI anyways.