Remove raw markdown output
I don't think it's ever been that useful. It was originally implemented
as kind of a test of what Hugo output formats were capable of. It waited
for the `ugly` option to get added in `outputFormats` and the novelty of
a new Hugo feature was part of the fun of it.

As mentioned in 7a65f92f53c3f40c33148bb88b4e1821edebee0f:

    I believe I originally added it because I thought it was somewhat
    hard to get from the "View history" commit list to the actual page
    but I feel like people who care enough to go looking for the source
    code will be willing to use GitHub to find the actual source file.